requirements.txt
Django==4.2.16
django-active-link==0.1.8
django-ckeditor==6.7.0
django-widget-tweaks
python-dotenv==1.0.0